modify homework lesson 5 add class Figure
