+++ Maven/src/main/java/lesson5/task2/Circle.java
@@ -1,9 +1,7 @@
 package lesson5.task2;
 
-public class Circle implements Shape, Fillable, Borderable {
+public class Circle extends Figure {
     private double radius;
-    private String fillColor;
-    private String borderColor;
 
     public Circle(double radius) {
         this.radius = radius;
@@ -15,27 +13,7 @@ public class Circle implements Shape, Fillable, Borderable {
 
     @Override
     public double calculateArea() {
-        return Math.PI * radius * radius;
-    }
-
-    @Override
-    public void setFillColor(String color) {
-        this.fillColor = color;
-    }
-
-    @Override
-    public String getFillColor() {
-        return fillColor;
-    }
-
-    @Override
-    public void setBorderColor(String color) {
-        this.borderColor = color;
-    }
-
-    @Override
-    public String getBorderColor() {
-        return borderColor;
+        return calculateCircleArea(radius);
     }
 
     public void printCharacteristics() {
